+++ src/Control/Exception.hs
@@ -2,3 +2,4 @@ module Control.Exception (
   module Base
 ) where
 import "base" Control.Exception as Base
+import GHC.Exception ()