+++ src/Control/Exception/Base.hs
@@ -2,3 +2,4 @@ module Control.Exception.Base (
   module Base
 ) where
 import "base" Control.Exception.Base as Base
+import GHC.Exception ()