+++ test/Control/ExceptionSpec.hs
@@ -14,3 +14,10 @@ spec = do
       -- The main purpose of this test is to check that `catch` is not
       -- ambiguated by Prelude.catch with older versions of base.
       (throwIO DivideByZero) `catch` (`shouldBe` DivideByZero)
+
+  describe "ErrorCall" $ do
+    it "has an Eq instance" $ do
+      ErrorCall "foo" `shouldBe` ErrorCall "foo"
+
+    it "has an Ord instance" $ do
+      ErrorCall "foo" `shouldSatisfy` (> ErrorCall "bar")