+++ CHANGES.markdown
@@ -3,6 +3,7 @@
  - provides access to `VERSION_base` and `MIN_VERSION_base` CPP macros (with
    `#include "base-compat.h"`)
  - Do not re-export `System.IO.Error.catch` from `Prelude` for `base` < 4.6.0
+ - Add `Eq`/`Ord` instance for `ErrorCall`
 
 ## Changes in 0.1.0
  - Remove getExecutablePath, it did not work with GHC < 7.2 (patches welcome!)